requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
absl-py=1.3.0=pypi_0
astunparse=1.6.3=pypi_0
blas=1.0=openblas
bottleneck=1.3.5=py39heec5a64_0
brotli=1.0.9=h1a28f6b_7
brotli-bin=1.0.9=h1a28f6b_7
c-ares=1.18.1=h1a28f6b_0
ca-certificates=2022.10.11=hca03da5_0
cachetools=5.2.0=pypi_0
cairo=1.16.0=h302bd0f_3
certifi=2022.12.7=py39hca03da5_0
charset-normalizer=2.1.1=pypi_0
contourpy=1.0.5=py39h525c30c_0
cycler=0.11.0=pyhd3eb1b0_0
eigen=3.3.7=h525c30c_1
expat=2.4.9=hc377ac9_0
fftw=3.3.9=h1a28f6b_1
flatbuffers=22.12.6=pypi_0
fontconfig=2.14.1=h6b8db82_1
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h1192e45_0
gast=0.4.0=pypi_0
gettext=0.21.0=h826f4ad_0
giflib=5.2.1=h1a28f6b_0
glib=2.69.1=h514c7bf_2
google-auth=2.15.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
graphite2=1.3.14=hc377ac9_1
grpcio=1.51.1=pypi_0
gst-plugins-base=1.14.1=hf0a386a_0
gstreamer=1.14.1=he09cfb7_0
h5py=3.7.0=pypi_0
harfbuzz=4.3.0=hb1b0ec1_0
hdf5=1.12.1=h160e8cb_2
icu=68.1=hc377ac9_0
idna=3.4=pypi_0
importlib-metadata=5.2.0=pypi_0
joblib=1.1.1=py39hca03da5_0
jpeg=9e=h1a28f6b_0
keras=2.11.0=pypi_0
kiwisolver=1.4.2=py39hc377ac9_0
krb5=1.19.2=h3b8d789_0
lcms2=2.12=hba8e193_0
lerc=3.0=hc377ac9_0
libbrotlicommon=1.0.9=h1a28f6b_7
libbrotlidec=1.0.9=h1a28f6b_7
libbrotlienc=1.0.9=h1a28f6b_7
libclang=14.0.6=pypi_0
libcurl=7.86.0=h0f1d93c_0
libcxx=14.0.6=h848a8c0_0
libdeflate=1.8=h1a28f6b_5
libedit=3.1.20221030=h80987f9_0
libev=4.33=h1a28f6b_1
libffi=3.4.2=hca03da5_6
libgfortran=5.0.0=11_3_0_hca03da5_28
libgfortran5=11.3.0=h009349e_28
libiconv=1.16=h1a28f6b_2
libllvm12=12.0.0=h12f7ac0_4
libnghttp2=1.46.0=h95c9599_0
libopenblas=0.3.21=h269037a_0
libpng=1.6.37=hb8d0fd4_0
libpq=12.9=h65cfe13_3
libssh2=1.10.0=hf27765b_0
libtiff=4.4.0=h2fd578a_2
libwebp=1.2.4=h68602c7_0
libwebp-base=1.2.4=h1a28f6b_0
libxml2=2.9.14=h8c5e841_0
libxslt=1.1.35=h9833966_0
llvm-openmp=14.0.6=hc6e5704_0
lz4-c=1.9.4=h313beb8_0
markdown=3.4.1=pypi_0
markupsafe=2.1.1=pypi_0
matplotlib=3.6.2=py39hca03da5_0
matplotlib-base=3.6.2=py39h8bbb115_0
munkres=1.1.4=py_0
ncurses=6.3=h1a28f6b_3
nspr=4.33=hc377ac9_0
nss=3.74=h142855e_0
numexpr=2.8.4=py39h79ee842_0
numpy=1.24.0=pypi_0
numpy-base=1.23.4=py39h90707a3_0
oauthlib=3.2.2=pypi_0
opencv=4.6.0=py39h8794c10_2
openssl=1.1.1s=h1a28f6b_0
opt-einsum=3.3.0=pypi_0
packaging=22.0=py39hca03da5_0
pandas=1.5.2=py39h78102c4_0
pcre=8.45=hc377ac9_0
pillow=9.3.0=py39hf4a492f_1
pip=22.3.1=py39hca03da5_0
pixman=0.40.0=h1a28f6b_0
protobuf=3.19.6=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pyparsing=3.0.9=py39hca03da5_0
python=3.9.15=hc0d8a6c_2
python-dateutil=2.8.2=pyhd3eb1b0_0
pytz=2022.7=py39hca03da5_0
qt-main=5.15.2=ha2d02b5_7
qt-webengine=5.15.9=h2903aaf_4
qtwebkit=5.212=h0f11f3c_4
readline=8.2=h1a28f6b_0
requests=2.28.1=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
scikit-learn=1.1.3=py39hc377ac9_0
scipy=1.9.3=py39h9d039d2_0
setuptools=65.5.0=py39hca03da5_0
six=1.16.0=pyhd3eb1b0_1
sklearn=0.0.post1=pypi_0
sqlite=3.40.0=h7a7dc30_0
tensorboard=2.11.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorflow-estimator=2.11.0=pypi_0
tensorflow-macos=2.11.0=pypi_0
termcolor=2.1.1=pypi_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=hb8d0fd4_0
tornado=6.2=py39h1a28f6b_0
typing-extensions=4.4.0=pypi_0
tzdata=2022g=h04d1e81_0
urllib3=1.26.13=pypi_0
werkzeug=2.2.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
wrapt=1.14.1=pypi_0
xz=5.2.8=h80987f9_0
zipp=3.11.0=pypi_0
zlib=1.2.13=h5a0b063_0
zstd=1.5.2=h8574219_0
